refactor: move overflow mixin to rendering dir
