+++ kraken/lib/src/rendering/overflow.dart
@@ -1,0 +1,265 @@
+/*
+ * Copyright (C) 2019-present Alibaba Inc. All rights reserved.
+ * Author: Kraken Team.
+ */
+
+import 'dart:math' as math;
+import 'package:flutter/rendering.dart';
+import 'package:kraken/rendering.dart';
+import 'package:kraken/css.dart';
+import 'package:kraken/gesture.dart';
+
+mixin RenderOverflowMixin on RenderBoxModelBase {
+  ScrollListener? scrollListener;
+  void Function(PointerEvent)? scrollablePointerListener;
+
+  void disposeScrollable() {
+    scrollListener = null;
+    scrollablePointerListener = null;
+    _scrollOffsetX = null;
+    _scrollOffsetY = null;
+    // Dispose clip layer.
+    _clipRRectLayer.layer = null;
+    _clipRectLayer.layer = null;
+  }
+
+  bool get clipX {
+    RenderBoxModel renderBoxModel = this as RenderBoxModel;
+    // Recycler layout not need repaintBoundary and scroll/pointer listeners,
+    // ignoring overflowX or overflowY sets, which handle it self.
+    if (renderBoxModel is RenderSliverListLayout) {
+      return false;
+    }
+
+    // The content of replaced elements is always trimmed to the content edge curve.
+    // https://www.w3.org/TR/css-backgrounds-3/#corner-clipping
+    List<Radius>? borderRadius = renderStyle.borderRadius;
+    if (renderBoxModel is RenderIntrinsic && borderRadius != null) {
+      return true;
+    }
+
+    // Overflow value other than 'visible' always need to clip content.
+    // https://www.w3.org/TR/css-overflow-3/#overflow-properties
+    CSSOverflowType effectiveOverflowX = renderStyle.effectiveOverflowX;
+    if (effectiveOverflowX != CSSOverflowType.visible) {
+      Size scrollableSize = renderBoxModel.scrollableSize;
+      Size scrollableViewportSize = renderBoxModel.scrollableViewportSize;
+      if (scrollableSize.width > scrollableViewportSize.width) {
+        return true;
+      }
+    }
+
+    return false;
+  }
+
+  bool get clipY {
+    RenderBoxModel renderBoxModel = this as RenderBoxModel;
+    // Recycler layout not need repaintBoundary and scroll/pointer listeners,
+    // ignoring overflowX or overflowY sets, which handle it self.
+    if (renderBoxModel is RenderSliverListLayout) {
+      return false;
+    }
+
+    // The content of replaced elements is always trimmed to the content edge curve.
+    // https://www.w3.org/TR/css-backgrounds-3/#corner-clipping
+    List<Radius>? borderRadius = renderStyle.borderRadius;
+    if (renderBoxModel is RenderIntrinsic && borderRadius != null) {
+      return true;
+    }
+
+    // Overflow value other than 'visible' always need to clip content.
+    // https://www.w3.org/TR/css-overflow-3/#overflow-properties
+    CSSOverflowType effectiveOverflowY = renderStyle.effectiveOverflowY;
+    if (effectiveOverflowY != CSSOverflowType.visible) {
+      Size scrollableSize = renderBoxModel.scrollableSize;
+      Size scrollableViewportSize = renderBoxModel.scrollableViewportSize;
+      if (scrollableSize.height > scrollableViewportSize.height) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  bool get enableScrollX {
+    CSSOverflowType effectiveOverflowX = renderStyle.effectiveOverflowX;
+    switch (effectiveOverflowX) {
+      case CSSOverflowType.hidden:
+      case CSSOverflowType.auto:
+      case CSSOverflowType.scroll:
+        return true;
+      case CSSOverflowType.clip:
+      case CSSOverflowType.visible:
+        return false;
+    }
+  }
+
+  bool get enableScrollY {
+    CSSOverflowType effectiveOverflowY = renderStyle.effectiveOverflowY;
+    switch (effectiveOverflowY) {
+      case CSSOverflowType.hidden:
+      case CSSOverflowType.auto:
+      case CSSOverflowType.scroll:
+        return true;
+      case CSSOverflowType.clip:
+      case CSSOverflowType.visible:
+        return false;
+    }
+  }
+
+  Size? _scrollableSize;
+  Size? _viewportSize;
+
+  ViewportOffset? get scrollOffsetX => _scrollOffsetX;
+  ViewportOffset? _scrollOffsetX;
+  set scrollOffsetX(ViewportOffset? value) {
+    if (value == _scrollOffsetX) return;
+    _scrollOffsetX?.removeListener(_scrollXListener);
+    _scrollOffsetX = value;
+    _scrollOffsetX?.addListener(_scrollXListener);
+    markNeedsLayout();
+  }
+
+  ViewportOffset? get scrollOffsetY => _scrollOffsetY;
+  ViewportOffset? _scrollOffsetY;
+  set scrollOffsetY(ViewportOffset? value) {
+    if (value == _scrollOffsetY) return;
+    _scrollOffsetY?.removeListener(_scrollYListener);
+    _scrollOffsetY = value;
+    _scrollOffsetY?.addListener(_scrollYListener);
+    markNeedsLayout();
+  }
+
+  void _scrollXListener() {
+    assert(scrollListener != null);
+    scrollListener!(scrollOffsetX!.pixels, AxisDirection.right);
+    markNeedsPaint();
+  }
+
+  void _scrollYListener() {
+    assert(scrollListener != null);
+    scrollListener!(scrollOffsetY!.pixels, AxisDirection.down);
+    markNeedsPaint();
+  }
+
+  void _setUpScrollX() {
+    _scrollOffsetX!.applyViewportDimension(_viewportSize!.width);
+    _scrollOffsetX!.applyContentDimensions(0.0, math.max(0.0, _scrollableSize!.width - _viewportSize!.width));
+  }
+
+  void _setUpScrollY() {
+    _scrollOffsetY!.applyViewportDimension(_viewportSize!.height);
+    _scrollOffsetY!.applyContentDimensions(0.0, math.max(0.0, _scrollableSize!.height - _viewportSize!.height));
+  }
+
+  void setUpOverflowScroller(Size scrollableSize, Size viewportSize) {
+    _scrollableSize = scrollableSize;
+    _viewportSize = viewportSize;
+    if (clipX && _scrollOffsetX != null) {
+      _setUpScrollX();
+    }
+
+    if (clipY && _scrollOffsetY != null) {
+      _setUpScrollY();
+    }
+  }
+
+  double get _paintOffsetX {
+    if (_scrollOffsetX == null) return 0.0;
+    return -_scrollOffsetX!.pixels;
+  }
+  double get _paintOffsetY {
+    if (_scrollOffsetY == null) return 0.0;
+    return -_scrollOffsetY!.pixels;
+  }
+
+  double get scrollTop {
+    if (_scrollOffsetY == null) return 0.0;
+    return _scrollOffsetY!.pixels;
+  }
+
+  double get scrollLeft {
+    if (_scrollOffsetX == null) return 0.0;
+    return _scrollOffsetX!.pixels;
+  }
+
+  bool _shouldClipAtPaintOffset(Offset paintOffset, Size childSize) {
+    return paintOffset < Offset.zero || !(Offset.zero & size).contains((paintOffset & childSize).bottomRight);
+  }
+
+  final LayerHandle<ClipRRectLayer> _clipRRectLayer = LayerHandle<ClipRRectLayer>();
+  final LayerHandle<ClipRectLayer> _clipRectLayer = LayerHandle<ClipRectLayer>();
+
+  void paintOverflow(PaintingContext context, Offset offset, EdgeInsets borderEdge, BoxDecoration? decoration, PaintingContextCallback callback) {
+    if (clipX == false && clipY == false) return callback(context, offset);
+    final double paintOffsetX = _paintOffsetX;
+    final double paintOffsetY = _paintOffsetY;
+    final Offset paintOffset = Offset(paintOffsetX, paintOffsetY);
+    // Overflow should not cover border
+    Rect clipRect = Offset(borderEdge.left, borderEdge.top) & Size(
+      size.width - borderEdge.right - borderEdge.left,
+      size.height - borderEdge.bottom - borderEdge.top,
+    );
+    if (_shouldClipAtPaintOffset(paintOffset, size)) {
+      // ignore: prefer_function_declarations_over_variables
+      PaintingContextCallback painter = (PaintingContext context, Offset offset) {
+        callback(context, offset + paintOffset);
+      };
+
+      // It needs to create new layer to clip children in case children has its own layer
+      // for all overflow value which is not visible (auto/scroll/hidden/clip).
+      bool _needsCompositing = true;
+
+      if (decoration != null && decoration.borderRadius != null) {
+        BorderRadius radius = decoration.borderRadius as BorderRadius;
+        RRect clipRRect = RRect.fromRectAndCorners(clipRect,
+          topLeft: radius.topLeft,
+          topRight: radius.topRight,
+          bottomLeft: radius.bottomLeft,
+          bottomRight: radius.bottomRight
+        );
+        _clipRRectLayer.layer = context.pushClipRRect(_needsCompositing, offset, clipRect, clipRRect, painter, oldLayer: _clipRRectLayer.layer);
+      } else {
+        _clipRectLayer.layer = context.pushClipRect(_needsCompositing, offset, clipRect, painter, oldLayer: _clipRectLayer.layer);
+      }
+    } else {
+      _clipRectLayer.layer = null;
+      _clipRRectLayer.layer = null;
+      callback(context, offset);
+    }
+  }
+
+  @override
+  double? computeDistanceToActualBaseline(TextBaseline baseline) {
+    double? result;
+    final BoxParentData? childParentData = parentData as BoxParentData?;
+    double? candidate = getDistanceToActualBaseline(baseline);
+    if (candidate != null) {
+      candidate += childParentData!.offset.dy;
+      if (result != null)
+        result = math.min(result, candidate);
+      else
+        result = candidate;
+    }
+    return result;
+  }
+
+  void applyOverflowPaintTransform(RenderBox child, Matrix4 transform) {
+    final Offset paintOffset = Offset(_paintOffsetX, _paintOffsetY);
+    transform.translate(paintOffset.dx, paintOffset.dy);
+  }
+
+  @override
+  Rect? describeApproximatePaintClip(RenderObject child) {
+    final Offset paintOffset = Offset(_paintOffsetX, _paintOffsetY);
+    if (_shouldClipAtPaintOffset(paintOffset, size)) return Offset.zero & size;
+    return null;
+  }
+
+  void debugOverflowProperties(DiagnosticPropertiesBuilder properties) {
+    if (_scrollableSize != null) properties.add(DiagnosticsProperty('scrollableSize', _scrollableSize));
+    if (_viewportSize != null) properties.add(DiagnosticsProperty('viewportSize', _viewportSize));
+    properties.add(DiagnosticsProperty('clipX', clipX));
+    properties.add(DiagnosticsProperty('clipY', clipY));
+  }
+}
+
